test(popover): migrate adjustment
